Improve exemple_5
 - splited into few files
 - directory for resources and classes
